Automated Detection Of Electrical Resistance Values Using Camera
the purpose of this project is to develop a system that can automatically detect the color bands on resistors and accurately calculate their resistance values. Measuring resistors manually is a time-consuming and error-prone process. In this project, we aim to solve this problem by combining deep learning and image processing techniques. Our system's portability and cost-effectiveness offer significant advantages, especially for industrial and educational applications